Refactoring to prepare for dynamic price scraping of multiple DEXs
- Set up a folder for all scraper files and moved the scraper file inot
this folder
- Created a file for the minswap scraping functionality and moved all of
the associated methods there.
- Updated the scrapeMinswapPrice method to call the method for
extracting the price from the innerText string we retrieve from the DEX,
as this is a minswap specific function.
- Created a scraperMethods file and an object that will be used to
access a DEX scraper method by using the DEX url as the key.
- Created a scraperUtils file and created an object that will be used to
acess any methods that are used across all DEX scraping actions.
- Extracted the code for navigating to a specified URL into it's own method and added it as a value to the scraperUtils object.
- Updated the scrapePrice method in the scraper file, so it now imports
the scraperUtils and scraperMethods.  Replaced the code for navigating
to a URL with the new scraperUtil method which does the same.
- Replaced the code for scraping prices with a dynamic call to the relevant scraperMethod by passing in the URL.
- Removed redundant code and methods to tidy up the scraper file.