Wingriders scraper method created
This method allows for scraping of pool prices from the wingriders dex

- Created a method which will try to find a pool and get the swap pair
values, divide the ADA value by the other currency value to get the price of 1 of that type of currency.
- Added a loop which checks if the pool was found.  If not, it will
navigate to the next page and check again. It repeats the process until
it has checked every page